backend/requirements.txt
fastapi
uvicorn
docxtpl
python-dotenv
requests
python-docx
google-generativeai
aiofiles
httpx
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
python-multipart
azure-cosmos
email-validator
